// src/types/Repo.ts
/**
 * Интерфейс, представляющий репозиторий на платформе GitHub.
 */
export interface Repo {
  /**
   * Уникальный идентификатор репозитория.
   */
  id: number;

  /**
   * Имя репозитория.
   */
  name: string;

  /**
   * Количество форков данного репозитория.
   */
  forks_count: number;

  /**
   * Основной язык программирования репозитория. Может быть `null`, если язык не указан.
   */
  language: string | null;

  /**
   * Дата последнего обновления репозитория в формате ISO 8601.
   */
  updated_at: string;

  /**
   * Количество звезд, полученных репозиторием.
   */
  stargazers_count: number;

  /**
   * Описание репозитория.
   */
  description: string;

  /**
   * Лицензия, под которой распространяется репозиторий.
   */
  license: RepoLicense;
}

/**
 * Интерфейс, представляющий лицензию репозитория.
 */
interface RepoLicense {
  /**
   * Уникальный ключ лицензии.
   */
  key: string;

  /**
   * Имя лицензии. Может быть `null`, если имя не указано.
   */
  name: string | null;

  /**
   * Node ID лицензии (уникальный идентификатор).
   */
  node_id: string;

  /**
   * SPDX идентификатор лицензии.
   */
  spdx_id: string;

  /**
   * URL, ссылающийся на страницу лицензии.
   */
  url: string;
}